Server: Add rate limit for screen login endpoint. Add StatusPages plugin Add screen routes (for logout screen and add styles for screen) Add image and price params for Category Add deletion files of deleted imageItem or foodItem

+++ Server/src/main/kotlin/routes/CategoryRoutes.kt
@@ -1,6 +1,7 @@
 package com.kalashnikovprojects.ufmserver.routes
 
 import com.kalashnikovprojects.ufmserver.adapters.eventbus.EventBus
+import com.kalashnikovprojects.ufmserver.adapters.filestorage.FileStorageAdapter
 import com.kalashnikovprojects.ufmserver.data.repository.CategoriesRepository
 import com.kalashnikovprojects.ufmserver.data.repository.FoodItemsCategoriesRepository
 import com.kalashnikovprojects.ufmserver.dto.Events
@@ -8,10 +9,14 @@ import com.kalashnikovprojects.ufmserver.dto.FoodItem
 import com.kalashnikovprojects.ufmserver.dto.NoIdCategory
 import com.kalashnikovprojects.ufmserver.dto.toCategory
 import io.ktor.http.HttpStatusCode
+import io.ktor.http.content.PartData
+import io.ktor.http.content.forEachPart
+import io.ktor.http.content.streamProvider
 import io.ktor.server.auth.authenticate
 import io.ktor.server.auth.jwt.JWTPrincipal
 import io.ktor.server.auth.principal
 import io.ktor.server.request.receive
+import io.ktor.server.request.receiveMultipart
 import io.ktor.server.response.respond
 import io.ktor.server.routing.*
 import org.koin.ktor.ext.inject
@@ -21,6 +26,8 @@ import kotlin.getValue
 fun Route.categoryRoutes() {
     val categoriesRepository by inject<CategoriesRepository>()
     val foodItemsCategoriesRepository by inject<FoodItemsCategoriesRepository>()
+
+    val fileStorageAdapter by inject<FileStorageAdapter>()
 
     val eventBus by inject<EventBus>()
 
@@ -97,16 +104,26 @@ fun Route.categoryRoutes() {
             val userId = principal!!.payload.getClaim("id").asInt()
 
             val request = call.receive<NoIdCategory>()
-
-            try {
-                val createdId = categoriesRepository.create(userId, request)
-                eventBus.getFlow(userId).emit(Events.AddCategoryEvent(
-                    element = request.toCategory(createdId)
-                ))
-                call.respond(HttpStatusCode.Created, createdId)
-            } catch (e: Exception) {
-                call.respond(HttpStatusCode.InternalServerError, e.localizedMessage ?: "Unknown error")
+            val multipart = call.receiveMultipart()
+            var fileBytes: ByteArray? = null
+            var fileName = ""
+            multipart.forEachPart { part ->
+                if (part is PartData.FileItem) {
+                    fileBytes = part.streamProvider().readBytes()
+                    fileName = part.originalFileName as String
+                }
+                part.dispose()
             }
+            fileBytes?.let { bytes ->
+                val imageUrl = fileStorageAdapter.saveFoodItemImage(bytes, fileName.substringAfterLast(".", ""))
+                request.imageUri = imageUrl
+            }
+
+            val createdId = categoriesRepository.create(userId, request)
+            eventBus.getFlow(userId).emit(Events.AddCategoryEvent(
+                element = request.toCategory(createdId)
+            ))
+            call.respond(HttpStatusCode.Created, createdId)
         }
 
         put("/categories/{id}") {
@@ -118,8 +135,22 @@ fun Route.categoryRoutes() {
                 call.respond(HttpStatusCode.BadRequest, "Invalid ID")
                 return@put
             }
-
             val request = call.receive<NoIdCategory>()
+
+            val multipart = call.receiveMultipart()
+            var fileBytes: ByteArray? = null
+            var fileName = ""
+            multipart.forEachPart { part ->
+                if (part is PartData.FileItem) {
+                    fileBytes = part.streamProvider().readBytes()
+                    fileName = part.originalFileName as String
+                }
+                part.dispose()
+            }
+            fileBytes?.let { bytes ->
+                val imageUrl = fileStorageAdapter.saveFoodItemImage(bytes, fileName.substringAfterLast(".", ""))
+                request.imageUri = imageUrl
+            }
             val isUpdated = categoriesRepository.updateById(userId, id, request)
 
             if (isUpdated) {
@@ -144,10 +175,20 @@ fun Route.categoryRoutes() {
                 call.respond(HttpStatusCode.BadRequest, "Invalid ID")
                 return@delete
             }
+            val categoryItem = categoriesRepository.getById(userId, id)
+            if (categoryItem == null) {
+                call.respond(HttpStatusCode.NotFound)
+                return@delete
+            }
 
             val isDeleted = categoriesRepository.deleteById(userId, id)
             if (isDeleted) {
-                Events.DeleteCategoryEvent(id)
+                categoryItem.imageUri?.let {
+                    fileStorageAdapter.deleteCategoryItemImage(categoryItem.imageUri)
+                }
+                eventBus.getFlow(userId).emit(
+                    Events.DeleteCategoryEvent(id)
+                )
                 call.respond(HttpStatusCode.OK)
             } else {
                 call.respond(HttpStatusCode.NotFound)
