+++ Server/src/main/kotlin/routes/ImageItemsRoutes.kt
@@ -57,32 +57,28 @@ fun Route.imageItemsRoutes() {
         post("/image-items") {
             val principal = call.principal<JWTPrincipal>()
             val userId = principal!!.payload.getClaim("id").asInt()
-            try {
-                val multipart = call.receiveMultipart()
-                var fileBytes: ByteArray? = null
-                var fileName = ""
-                multipart.forEachPart { part ->
-                    if (part is PartData.FileItem) {
-                        fileBytes = part.streamProvider().readBytes()
-                        fileName = part.originalFileName as String
-                    }
-                    part.dispose()
+            val multipart = call.receiveMultipart()
+            var fileBytes: ByteArray? = null
+            var fileName = ""
+            multipart.forEachPart { part ->
+                if (part is PartData.FileItem) {
+                    fileBytes = part.streamProvider().readBytes()
+                    fileName = part.originalFileName as String
                 }
-                if (fileBytes == null) {
-                    call.respond(HttpStatusCode.BadRequest, "No file uploaded")
-                    return@post
-                }
-                val imageUrl = fileStorageAdapter.saveFoodItemImage(fileBytes, fileName.substringAfterLast(".", ""))
-                val image = NoIdImageItem(imageUrl)
-
-                val createdId = imageItemsRepository.create(userId, image)
-                eventBus.getFlow(userId).emit(Events.AddImageEvent(
-                    element = image.toImageItem(createdId)
-                ))
-                call.respond(HttpStatusCode.Created, image.toImageItem(createdId))
-            } catch (e: Exception) {
-                call.respond(HttpStatusCode.InternalServerError)
+                part.dispose()
             }
+            if (fileBytes == null) {
+                call.respond(HttpStatusCode.BadRequest, "No file uploaded")
+                return@post
+            }
+            val imageUrl = fileStorageAdapter.saveImageItemImage(fileBytes, fileName.substringAfterLast(".", ""))
+            val image = NoIdImageItem(imageUrl)
+
+            val createdId = imageItemsRepository.create(userId, image)
+            eventBus.getFlow(userId).emit(Events.AddImageEvent(
+                element = image.toImageItem(createdId)
+            ))
+            call.respond(HttpStatusCode.Created, image.toImageItem(createdId))
         }
 
         delete("/image-items/{id}") {
@@ -94,9 +90,14 @@ fun Route.imageItemsRoutes() {
                 call.respond(HttpStatusCode.BadRequest, "Invalid ID")
                 return@delete
             }
-
+            val imageItem = imageItemsRepository.getById(userId, id)
+            if (imageItem == null) {
+                call.respond(HttpStatusCode.NotFound)
+                return@delete
+            }
             val isDeleted = imageItemsRepository.deleteById(userId, id)
             if (isDeleted) {
+                fileStorageAdapter.deleteImageItemImage(imageItem.imageUri)
                 eventBus.getFlow(userId).emit(Events.DeleteImageEvent(
                     id,
                 ))
